Refactoring as a concern.
